5 - Quinto/Trabajo Final/Trabajo Final/Código/TF/requirements.txt
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
bokeh==1.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
construct==2.10.56
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-bootstrap4==1.1.1
django-extensions==2.2.8
django-filter==2.2.0
django-tables2==2.2.1
entrypoints==0.3
graphviz==0.13.2
gunicorn==20.0.4
heroku==0.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kiwisolver==1.1.0
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
pickleshare==0.7.5
Pillow==7.0.0
pip-autoremove==0.9.1
prometheus-client==0.6.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyct==0.4.6
pydot==1.4.1
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
pyviz-comms==0.7.3
qtconsole==4.6.0
requests==2.23.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
static3==0.7.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1